Add files via upload
This coursework project represents a fundamental ASP.NET web application integrated with a database through Entity Framework. The application facilitates administrative tasks, allowing the addition and deletion of data in the form of text. The simplicity of the current text-based functionality sets the stage for potential expansion, enabling the inclusion of multimedia elements such as pictures and videos.